Add Unity 3-cushion billiards game with UaaL integration
- New Game tab (Play) with MAUI → Unity bridge (Android/iOS/Windows)
- Unity project: billiard table, 3 balls, cushions, legs, camera controller
- BilliardSceneBuilder editor script for automated scene setup
- Innings field changed to double (per-game average, e.g. 0.400)
- Stats fix: removed LastMatchResult, fixed GetStatsAsync query
- App name changed to BilliardIQ, added Game page localization
- Unity .aar embedded with Bind=false, manifest conflicts resolved

+++ BilliardIQ.Mobile/Models/Game.cs
@@ -10,7 +10,7 @@ public class Game
     public int PlayerScore { get; set; }
     public int OpponentScore { get; set; }
     public int HighestRun { get; set; }
-    public int Innings { get; set; }
+    public double Innings { get; set; }
     public string? Notes { get; set; }
     public string? ScoreboardPhotoPath { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
